pyproject.toml
[tool.poetry]
name = "imobiliariaapi"
version = "0.1.0"
description = ""
authors = ["felipe_santos"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.2"
python-dotenv = "^1.0.0"
djangorestframework = "^3.14.0"
mysqlclient = "^2.2.0"
djangorestframework-simplejwt = "^5.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
